requirements.txt
alembic==1.7.7
aniso8601==9.0.1
attrs==21.4.0
click==8.1.3
DateTime==4.4
flake8==4.0.1
Flask==2.1.2
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
flask-restplus==0.13.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
jsonpickle==2.1.0
jsonschema==4.4.0
Mako==1.2.0
MarkupSafe==2.1.1
mccabe==0.6.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pyrsistent==0.18.1
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.36
uuid==1.30
Werkzeug==2.1.2
zipp==3.8.0
zope.interface==5.4.0